New DropDown Menu Feature
I added a drop down menu that autocompletes the city name that you are trying to enter. You can click any of the boxes that pop up and it will display that locations weather as well as change the search bar to display the name of the city that you just clicked.